Add TLS (https) support
Introduces a new customizable boolean variable, http-twiddle-tls, which defaults
to `nil' to maintain the old behavior.

When set to `t' the network connection (as created with `open-network-stream')
will be of type "tls" insteaf of "plain", and the default port, when not
specified in the Host: header, will be 443 instead of 80.